refactor(api): remove provider routines service barrel
