--- api/app/src/services/provider-routines/index.ts
@@ -1,19 +1,0 @@
-export { callProviderRoutine } from "./call";
-export type {
-  ConnectorProviderRoutineAdapter,
-  ConnectorProviderRoutineTool,
-  LinearProviderRoutineAdapter,
-  ProviderRoutineActor,
-  ProviderRoutineServiceAdapters,
-  ProviderRoutineServiceContext,
-  ProviderRoutineServiceLog,
-  ProviderRoutineSource,
-} from "./context";
-export { ProviderRoutineError, providerRoutineError } from "./errors";
-export { findProviderRoutines } from "./find";
-export { defaultLinearProviderRoutineAdapter } from "./linear";
-export {
-  classifyLinearRoutine,
-  classifyRoutine,
-  hasRoutineScope,
-} from "./policy";
